separated dashboard code into component files
